Add options for disabling S3/DynamoDB access
Adds command line options for disabling the default permissions on S3 or
DynamoDB. For better security the API should only have access to the specific
resources it needs. The easiest way of doing that is probably to attach custom
policies to the API created role.

The default option is still full rights to all resources.